refactor(core): validation exception reorganization
