+++ packages/core/lib/validator/config.validator.ts
@@ -1,7 +1,7 @@
 import { validateSync } from 'class-validator';
 import { Type } from '../utils/type.interface';
-import { ConfigValidationException } from './config.validation.exception';
-import { ConfigValidationExceptionOptions } from './config.validation.exception.options';
+import { ConfigValidationException } from './exception/config.validation.exception';
+import { ConfigValidationExceptionOptions } from './exception/config.validation.exception.options';
 
 export class ConfigValidator {
   validate(configs: object[]) {
