+++ packages/core/lib/validator/exception/config.validation.exception.options.ts
@@ -1,5 +1,5 @@
 import { ValidationError } from 'class-validator';
-import { Type } from '../utils/type.interface';
+import { Type } from '../../utils/type.interface';
 
 export interface ConfigValidationExceptionOptions {
   template: Type;
