+++ packages/core/lib/validator/exception/config.validation.exception.ts
@@ -1,10 +1,17 @@
 import { ValidationError } from 'class-validator';
 import { getBorderCharacters, SpanningCellConfig, table } from 'table';
-import { PROPERTIES_MAPPING_METADATA } from '../loader/constants';
-import { PropertiesMapping, PropertySource, PropertyTarget } from '../loader/types';
+import { PROPERTIES_MAPPING_METADATA } from '../../loader/constants';
+import { PropertiesMapping, PropertySource, PropertyTarget } from '../../loader/types';
 import { ConfigValidationExceptionOptions } from './config.validation.exception.options';
 
-interface ValidationReportRow {
+const headers = ['Template', 'Property', 'Source', 'Current Value', 'Failed constraints'];
+
+// interface TemplateValidationErrors {
+//   template: ClassConstructor<any>;
+//   errors: ValidationError[];
+// }
+
+interface ValidationError {
   propertyTarget: PropertyTarget;
   propertySource?: PropertySource;
   failedConstraints: string[];
@@ -15,14 +22,13 @@ export class ConfigValidationException extends Error {
   constructor(failedValidations: ConfigValidationExceptionOptions[]) {
     super();
     this.name = ConfigValidationException.name;
-    this.message = '\n' + this.formatMessage(failedValidations);
+
+    const tableRows = this.formatTableRows(failedValidations);
+    this.message = '\n' + this.formatMessage(tableRows);
   }
 
-  private formatMessage(failedValidations: ConfigValidationExceptionOptions[]) {
-    const tableRows = this.formatTableRows(failedValidations);
-    const header = ['Template', 'Property', 'Source', 'Current Value', 'Failed constraints'];
-    const tableData = [header, ...tableRows.flatMap(({ templateTableData }) => templateTableData)];
-
+  private formatMessage(tableRows: { templateTableData: any[][]; spanningCells: SpanningCellConfig }[]) {
+    const tableData = [headers, ...tableRows.flatMap(({ templateTableData }) => templateTableData)];
     return table(tableData, {
       columns: [{ alignment: 'left', width: 16 }],
       spanningCells: tableRows.map(({ spanningCells }) => spanningCells),
@@ -36,7 +42,7 @@ export class ConfigValidationException extends Error {
       const templateTableData = this.formatExceptionsReport(failedValidation.errors).map((row) => [
         '',
         row.propertyTarget,
-        row.propertySource || '',
+        row.propertySource ?? '',
         row.currentValue,
         row.failedConstraints,
       ]);
@@ -53,8 +59,8 @@ export class ConfigValidationException extends Error {
     });
   }
 
-  private formatExceptionsReport(errors: ValidationError[], parentPath = ''): ValidationReportRow[] {
-    return errors.flatMap<ValidationReportRow>((error) => {
+  private formatExceptionsReport(errors: ValidationError[], parentPath = ''): ValidationError[] {
+    return errors.flatMap<ValidationError>((error) => {
       const propertyTarget = parentPath + error.property;
       if (error.children && error.children.length > 0) {
         return this.formatExceptionsReport(error.children, `${propertyTarget}.`);
