+++ packages/core/lib/validator/exception/index.ts
@@ -1,0 +1,1 @@
+export * from './config.validation.exception';
