+++ packages/core/lib/validator/index.ts
@@ -1,1 +1,1 @@
-export * from './config.validation.exception';
+export * from './exception';
